Refactor formatter functions into DefaultFormatters
